+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.5.0] - 2024-07-24
+
+## Features
+
+- merged family history of diseases from the survey data (e.g., Y/N)
+- added family history of cancer as a feature (e.g., Y/N)
+
 # [0.4.0] - 2024-07-23
 
 - Changed model optimization to use the optuna package